+++ rs/replica/src/setup_ic_stack.rs
@@ -32,10 +32,13 @@ use ic_types::{
     artifact::UnvalidatedArtifactMutation,
     consensus::{CatchUpPackage, HasHeight},
     messages::SignedIngress,
-    Height, NodeId, SubnetId,
+    Height, NodeId, PrincipalId, SubnetId,
 };
 use ic_xnet_payload_builder::XNetPayloadBuilderImpl;
-use std::sync::{Arc, RwLock};
+use std::{
+    str::FromStr,
+    sync::{Arc, RwLock},
+};
 use tokio::sync::{
     mpsc::{channel, UnboundedSender},
     watch, OnceCell,
@@ -45,8 +48,24 @@ use tokio::sync::{
 /// the message id and height of messages that complete execution.
 const COMPLETED_EXECUTION_MESSAGES_BUFFER_SIZE: usize = 10_000;
 
-/// True if the node should serve synchronous responses to v3 update calls.
-const ENABLE_SYNCHRONOUS_CALL_V3: bool = false;
+/// The subnets that should not serve synchronous responses to v3 update calls.
+/// The list contains all system subnets.
+const SUBNETS_WITH_DISABLED_SYNCHRONOUS_CALL_V3: [&str; 1] =
+    ["tdb26-jop6k-aogll-7ltgs-eruif-6kk7m-qpktf-gdiqx-mxtrf-vb5e6-eqe"];
+
+/// Returns true if the subnet is whitelisted to serve synchronous responses to v3
+/// update calls.
+fn enable_synchronous_call_handler_for_v3_endpoint(subnet_id: &SubnetId) -> bool {
+    let subnet_is_in_disabled_list =
+        SUBNETS_WITH_DISABLED_SYNCHRONOUS_CALL_V3
+            .iter()
+            .any(|s| match PrincipalId::from_str(s) {
+                Ok(principal_id) => SubnetId::from(principal_id) == *subnet_id,
+                Err(_) => false,
+            });
+
+    !subnet_is_in_disabled_list
+}
 
 /// Create the consensus pool directory (if none exists)
 fn create_consensus_pool_dir(config: &Config) {
@@ -354,7 +373,7 @@ pub fn construct_ic_stack(
         tracing_handle,
         max_certified_height_rx,
         finalized_ingress_height_rx,
-        ENABLE_SYNCHRONOUS_CALL_V3,
+        enable_synchronous_call_handler_for_v3_endpoint(&subnet_id),
     );
 
     Ok((